+++ src/Console/Command/RunTestsCommand.php
@@ -46,6 +46,7 @@ class RunTestsCommand extends Command
     const OPTION_PATTERN = 'pattern';
     const OPTION_GROUP = 'group';
     const OPTION_EXCLUDE_GROUP = 'exclude-group';
+    const OPTION_FILTER = 'filter';
     const OPTION_PUBLISH_RESULTS = 'publish-results';
 
     /**
@@ -126,6 +127,12 @@ class RunTestsCommand extends Command
                 null,
                 InputOption::VALUE_REQUIRED | InputOption::VALUE_IS_ARRAY,
                 'Exclude testcases with specified @group from being run'
+            )
+            ->addOption(
+                self::OPTION_FILTER,
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Run only testcases/tests with name matching this filter'
             )
             ->addOption(
                 self::OPTION_PUBLISH_RESULTS,
@@ -246,7 +253,8 @@ class RunTestsCommand extends Command
         $processSet = $processSetCreator->createFromFiles(
             $files,
             $input->getOption(self::OPTION_GROUP),
-            $input->getOption(self::OPTION_EXCLUDE_GROUP)
+            $input->getOption(self::OPTION_EXCLUDE_GROUP),
+            $input->getOption(self::OPTION_FILTER)
         );
 
         if (!count($processSet)) {